[#83] refactor: subscriber 를 presentation 으로 이동
